chore: rename modules by role and scrub private-dataset specifics
Behavior-neutral cleanup (C-3 + C-8): pure renames, re-exports, and
comment edits. No detection logic, no constant values, and no public
API path changed.

C-3 — rename same-name-different-role modules by role:
- projective-grid: shared/recovery.rs -> shared/recovery_schedule.rs
  (the geometry-only recovery *schedule*, entry `run_schedule`).
- projective-grid: shared/validate/recovery.rs ->
  shared/validate/wrong_label_filters.rs (the wrong-label *drop
  filters*: `topological_wrong_label_drops`, the frontier-kink filter,
  `largest_cardinally_connected_component`).
- charuco: detector/corner_validation.rs -> detector/corner_refit.rs
  (the homography corner refit; internal `pub(crate)`).
- charuco: validation.rs -> link_check.rs (the public marker-corner
  linkage check). Crate-root re-exports of every item name are
  preserved unchanged, and a deprecated `pub use crate::link_check as
  validation;` keeps `calib_targets_charuco::validation::*` resolving.

All in-workspace importers and intra-doc links were updated to the new
paths (no stale aliases left in `shared`); function/type names are
unchanged.

C-8 — scrub private-dataset specifics from charuco source comments:
Rewrote default-constant comments in detector/params.rs that cited
concrete private-dataset board sizes, private-frame dictionary
families, frame counts, and upscale factors into general statements
about the internal regression sets. Comments only; constant values
untouched.

Gates: cargo fmt --all --check, clippy --workspace --all-targets
-D warnings, test --workspace, and doc --workspace --no-deps all pass
(zero warnings).
