+++ crates/calib-targets-charuco/src/link_check.rs
